Fix alignment of debug lines
I tried to fix this for Milestone 1, but it was causing a strange crash in tinyECS. The issue was that we are looping over the Motion components/entities while adding stuff to the Motion registry (see DebugSystem::createLine). To fix that crash, I just cache the value of entities.size() before the loop. Since tinyECS adds new components/entities to the back of the list, the list can grow bigger and the loop won't be affected as long as we use the initial value of entities.size().